GLES: Fix OOB access of labelBlock.
This was being incorrectly sized by the number of functions, not the highest label in use.

While investigating this, I've realized that the sanity checks to ensure there are no dead functions was never going to fire as the function list was built from the list of called functions. Instead I've changed the function scanning pass to look for labels starting a LABEL-RET pair.

Bug: b/125183107
Change-Id: Ic921097ed42a96b52f1ab7c9590c02fb3552b565