Add -Wno-implicit-fallthrough
This commit adds `-Wno-implicit-fallthrough` to disable implicit
fallthrough warnings.  This commit is required because the latest
LLVM/Clang toolchain emits more warnings.  Since `third_party/llvm-7.0`
is an upstream project, we don't want to locally patch the code.  Thus,
we are disabling implicit fallthrough warnings before it has been fixed
in the upstream.

Bug: b/115344057
Test: Build libGLESv2_swiftshader for Android
Change-Id: Ibde53e030ba66af20ba5281b33a684350ef4b35f