Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I'm not actually against this change, but this issue is also addressed by https://swiftshader-review.googlesource.com/25348 which already landed downstream: https://android-review.googlesource.com/912015 and originates from upstream LLVM changes. So I think we should avoid discrepancy even though disabling the warning does seem like a fine solution as well.
> 
> Cody, do you intend to land your change in our upstream repo or do you have second thoughts about it?

I still intended to land that change, it sounded like you wanted to hear feedback from rnk@ on it first.

It looks like this would still be a possible change from the Android side, as even with the eventual transition from Android.mk to Android.bp there are plenty of Android.bp files that exist now with "-Wno-implicit-fallthrough" so this could survive that.

Patch-set: 1
CC: Gerrit User 7153 <7153@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>