Update patch set 1
Patch Set 1: Code-Review-1

Thanks for the proposal, but this type of construct generally complicates readability, code navigation, and debugging. There has to be a significant upside to using this for it to be considered, which is not readily apparent to me. Happy to take a closer look if you can point it out.

Note that SPIR-V was explicitly designed to "enable a single simple pass through a module to transform it". Splitting off the input/prolog and output/epilog makes sense to me, but I fear that anything more generic would just complicate things.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=-1