Allow the Blitter to clear and blit vk::Image objects directly
The sw::Surface object was the intermediate representation of an
image used between the vk::Image and the Blitter. This cl removes
the need for an intermediate representation by having the Blitter
use the vk::Image directly.

There should be no regression for the clear and blit tests.

Bug b/126883332

Change-Id: Icbc15470e3ad112ed78f4f62d6d82c66e3e37a20