Update patch set 14
Patch Set 14: Code-Review+2

Sorry for the delay. I finally wrapped my head around the confusion in the child CLs.

For this CL specifically, since SPIR-V is extremely close to LLVM's SSA and memory model, and arrays correspond to alloca'd memory on the stack it has to be perfectly valid to take the address of an element.

Unlike for basic types (scalars, vectors, and pointers), we can't represent them as 'registers' so it doesn't clash with the direct-to-SSA idea for lvalues, if we still wanted to explore that direction. In other words this operator& is still very useful and reasonable even if we remove operator& from lvalues (and we can probably keep the latter by promoting 'registers' back to stack memory).

Patch-set: 14
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+2