Update patch set 5
Patch Set 5:

> Patch Set 3: Code-Review-1
> 
> I do not intend to ever land this, or for this to be reviewed.
> It was simply to check what tests passed after the compute shader change.

I was wondering about that, but the commit message looked too clean. ;-) Also the increase in run time is negligible so I'd still be OK with this especially since you'll be focusing on compute for the next little while and much of this will be made to pass.

Again, your call when/if you want to land this.

Patch-set: 5
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>