Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+2
> > 
> > Please mention the test that this passes in the commit message.
> 
> I'm going to look into getting regres to do this sort of thing.
> 
> There's about 15 scattered tests that use SMod, and manually testing these before and after is extremely laborious for an opcode we know we need implemented.
> 
> I'll try to get regres to scan the CL description for something like 'Tests: *smod*' and have it pull in all of those tests that match the wildcards.

The important bit to me is knowing which test provides coverage of this code. Just one suffices if it's non trivial to get the whole set, but preferably the most isolated one (a test with smod in the name is much more interesting than one which tests e.g. matrices but happens to use smod).

The value is that if we touch the code again, we can look at the 'blame' history to quickly find which tests we can run manually quickly to aim not to break during the refactoring. Will be a significant time saver towards the end of the project and future development/maintenance.

Patch-set: 2
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>