Regres: Don't attempt to build failing changes forever.
If a test failed to build, an error was logged to stdout, but the error was not posted to the change. This meant the change would be picked up by regres again, and the process would repeat ad-infinitum.

Posting of the build error used to work, but was broken by bfaf4e825b.

Change-Id: I1e59f0d2e5ee26eb002aa2c596dc5491e48a9c87