Copy build artifacts to CMAKE_BINARY_DIR.
Currently when the build executes it copies things like .so files and
the vulkan .icd file to the CMAKE_CURRENT_SOURCE_DIR/out/<platform>
directory. This is fine in a normal swiftshader build but if you embed
swiftshader into another project this ends up being something like:
third_party/swiftshader/out/<platform> which is not what is desired.

This CL changes the files to get written to the CMAKE_BINARY_DIR. This
means it will go into the directory form which cmake was run. This means
for a build setup as per the README, build/<platform>.

This also allows generating different .icd and .so files for Release and
Debug if you setup different build directories.

Bug: b/123752208
Change-Id: If2e8c70ca2d35351c02f7bbe56ff77a04d538373