Update patch set 1
Patch Set 1:

Oh, didn't notice before that Regres had already run, with no failures. Does make some sense if we're already using it correctly as a fully exclusive lock in most places, and allowing the same thread to lock it again in the odd case wouldn't hurt.

Still, this is a big semantical change for the class, and I suspect it makes things more expensive. If a recursive lock is needed we should have a separate implementation for it. But first we need to get to the bottom of whether that's what's really needed.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>