SpirvShader: Minor changes for ASSERTs
sw::Intermediate would only clear the contents to zero in debug builds, but would always validate that these were nullptr in release (ASSERT still warns in release).
Given the cost nullifying this memory is negligable in comparison to the actual LLVM JIT, always clear.

Changed a bunch of ASSERT()s to ASSERT_MSG() where the additional information is useful.

Replaced a few remaining calls to assert() with ASSERT()

Bug: b/127433389
Change-Id: Ifac89ca061bf7d61ff7d0de1792eeda18fad275c