Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Note: dEQP-VK.pipeline.multisample.raster_samples.samples_4.primitive_point currently fails, as SwiftShader only produces 2 different colors, and the test requires 3 different colors in the result. All other dEQP-VK.pipeline.multisample.raster_samples.samples_4.* tests produce acceptable results and pass.
> 
> Do you think we're doing something wrong, or is this test too fussy?

I'll investigate this later, but the rendertargets are tiny and points are a single pixel, so it's not impossible that multisampling only generates 2 colors in this scenario. I'll have to take a look at the result, as the output in Cherry shows the result when the test passes, but not when the test fails :( I'll temporarily remove the 3 colors requirement to look at the result.

Patch-set: 1
Reviewer: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>