Remove the lock usage from eglSwapBuffers.
This lock is hard to reason about because the implementation of
eglSwapBuffers eventually goes into Android code, which then calls
additional methods on SwiftShader.

This change is an alternative to https://swiftshader-review.googlesource.com/c/SwiftShader/+/27408
The methods Android happens to call at the moment are
eglClientWaitSync[KHR] and eglDestroySync[KHR], so it would also be
possible to remove the locks from those methods instead and remove it
from eglSwapBuffers. This is a more conservative change, as it does not
make assumptions about which SwiftShader methods Android will attempt to
call in its graphics implementation.

Bug: b/124530765
Test: atest CtsGraphicsTestCases:android.graphics.cts.BitmapTest#testDrawingHardwareBitmapNotLeaking -- --abi x86
Signed-off-by: Cody Schuffelen <schuffelen@google.com>
Change-Id: Id0746e131253539a97d964cee390a3c3f3836a1e