Use a C array for temporary load results
Array<> emits GEP instructions on every access, which aren't necessary.
Note that these temporaries are necessary because the If/Else blocks
blocks both write to them, while the destination intermediate object
is an rvalue which can only be initialized once (SSA).

Also EmitAccessChain was moved to match the declaration order.

Bug b/128539387

Change-Id: I726fb0fd28b0a19a61e8759679e30bc699f3279d