Update patch set 12
Patch Set 12:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-1
> > 
> > This seems useful to me. I'm a huge fan of being able to check out a repo and immediately work be able to work on it. What part is user-specific? Can we separate it?
> 
> Paths to dEQP, or any other app you want to debug swiftshader with. The type of debugger you want to use. I have both GDB and LLDB variants of unit tests and dEQP. I have two versions of dEQP. I have different rules for attaching vs launching.

I'm fine with all of that being checked in, as long as it makes reasonable assumptions about the workstation setup that others can duplicate. The goal of having checked in project files is to make it easy to get started working on a fresh checkout. Many other projects make assumptions about projects being checked out side-by-side in a specific directory, if you want to run/build something that has a dependency on both.

> I literally have to stash and unstash these files every time I rebase, and have to maintain 3 different versions for 3 different machines.

Is it feasible to merge these three versions into one, or is it impossible to make launch.json have platform-specific parts?

> I looked into extracting external paths, and I am not aware of any way to do this.
> 
> > I'm a huge fan of being able to check out a repo and immediately work be able to work on it
> 
> You can't even do that right now. All of the entries in lanuch.json currently point to paths that are not typical of our build "${workspaceRoot}/debug/...". It doesn't seem like anyone has tried using these in a long time.

It's been a while, but I used it on my ARM64 Jetson Xavier. Is it in the way of anything? Isn't each configuration independent so you can just select the one(s) that works for you?

> Setting up a new launch file is not hard - when the file does not exist VSCode auto fills the file with 'insert here' tokens. As the file isn't checked in, this is a one-time-cost for a new team member, and they can put whatever paths / tools they like in there.

I just want to avoid having to put the file somewhere other than the repo. Renaming it is a reasonable compromise but I'd also like to avoid having everyone check in their own version of it. Ideally the main one can reasonably work for everybody.

Patch-set: 12