Pass descriptor sets to SPIR-V compilation
The at draw time currently bound descriptors are passed down to the
SPIR-V code generation so that we can read the sampler parameters and
image view parameters.

An important exception is compute shaders, which are currently
compiled before the invoke calls.

Also still TODO is taking the parameters into account during routine
lookup in the cache.

This change also eliminates the now unused 'enableMask' from
PixelProgram and VertexProgram.

Bug b/129523279

Change-Id: I1ea75ec5a7b83783a1efacb238143a419afa5bda