Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> The only Vulkan code change since the last test list result update is https://swiftshader-review.googlesource.com/c/SwiftShader/+/28890
> 
> Alexis, could you try to repro some of these failures both before and after your change?

Note: these crashes wouldn't have shown up in a presubmit Regres run unless they were explicitly added (via 'Test:') as they were not already passing (all appear to be UNIMPLEMENTED -> CRASH).

This transition is not the end of the world (should be easy enough to diagnose with a debugger later), I was just curious if this was known/expected or not.

Patch-set: 1