Swiftshader to reject large arrays of uniforms
Swiftshader's GLES compiler ran into an OOM issue when shaders tried to
allocate large uniform arrays. Swiftshader only checked the size of the
array after it allocated the memory for the uniforms. So if someone
allocated an array of INT_MAX floats, then the compiler would crash.

The GLES compiler now checks prior to allocation that an array is not
larger than GL_MAX_VERTEX_UNIFORM_VECTORS and the fragment equivalent.

unittests.cpp has a new test that makes sure we're failing to compile
in these tests.

Bug chromium:872321
Change-Id: I0c2c857f138c40d25c637fda89f9c5b6b1a9d64c