Update patch set 1
Patch Set 1:

How about having a const extern bool for every Reactor 'feature' that is set accordingly in LLVMReactor.cpp and SubzeroReactor.cpp? That way it doesn't complicate our various build files and the code to assert on using a missing feature would be more elegant than when using the preprocessor.

Patch-set: 1
CC: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>