WIP: ETC2 Image decompression and sampling
The ETC2 decoder already existed in SwiftShader, so this cl hooks it
up in vk::Image. In order to be transparent to the user, any time a
compressed image is created, another image is simultaneously created,
which will eventually contain the decompressed image.

The only thing that's missing is a call to
VkImageView::prepareForSampling() before sampling from that image.

Change-Id: I8bfdccccd9cad30e5b707ba82aac2b581ec2a90e