Implement OpEmitImageSampleExplicitLod
Emit a call to a different trampoline function for generating the
sampler code, which unifies the implementation with
OpEmitImageSampleExplicitLod's by using a common EmitImageSample which
takes a sampler 'method' enum. This is then passed at shader execution
time to emitSamplerFunction.

The lod parameter is parsed from the instruction stream in
EmitImageSample. Other (optional) image operands are left unimplemented
for now.

Bug: b/129523279
Test: dEQP-VK.binding_model.shader_access.primary_cmd_buf.combined_image_sampler_mutable.fragment.single_descriptor.2d
Change-Id: I265b81d953fe5a0496d029704a0f5eeff4229823