SpirvShader: Rework debug assertions
Add src/Pipeline/SpirvUnsupported.hpp. This has a few purposes:
(1) It keeps long stirngs out of cpp files.
(2) It keeps the descriptions consistent for a given feature.
(3) It lets us see all the SPIRV unsupported features in one place (and those we're not handling).
(4) When we implement a feature, we have a simple way to find all the places we need to fix up.

Use the new UNSUPPORTED() macro where possible.

Switch to using UNREACHABLE() over UNIMPLEMENTED() for switch cases that do not exist according to the spec, or for cases that should be handled by validation in an earlier stage.

Bug: b/131243109
Change-Id: Id9970b64cf248468c508e65366bc576cb03e8e58