CMakeLists: Fix REACTOR_EMIT_DEBUG_INFO builds
ENABLE_RR_DEBUG_INFO was not being passed to the ReactorLLVM target.
This introduced a subtle calling mismatch as the RValue<T> copy constructor is conditionally added only when ENABLE_RR_DEBUG_INFO is defined.
This would cause parameters to go out of sync for inlined functions (not entirely sure I understand the underlying cause here), leading to hilarity.

This change also moves "-fno-exceptions" back to the global cpp_flags. This is a true global setting and should not have been moved to SWIFTSHADER_COMPILE_OPTIONS.

Change-Id: Ib20777f1ee4b443358ada667902f9a6a1ac3513a