Fix for VK_DESCRIPTOR_TYPE_SAMPLER and VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE
The default memcpy in DescriptorSetLayout::WriteDescriptorSet was
wrong for these types and causing crashes, because we could attempt
to read out of bounds of the source data. For example, if the type
is VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE, the external descriptor set is
a VkDescriptorImageInfo which is much smaller than the internal
representation, which is a SampledImageDescriptor.

Bug b/123244275

Change-Id: I3fe89c72b1f6d5dc48037105fb3b1f0eba1a8ade
Tests: dEQP-VK.binding_model.shader_access.*