VkDescriptorSetLayout: Fixes / seatbelts for memory derefs
(1) Add GetDescriptorStride(), change GetDescriptorSize() back to unaligned.
(2) Change getOffsetPointer() to (optionally) return both the size and stride.
(3) memcpy using a size of typeSize and offset with typeStride. Prevents OOB reads.
(4) Align the result of getDescriptorSetAllocationSize to the alignment of DescriptorSet. This will currently be a no-op as each descriptor is 16 byte aligned anyway, but this is correct for any future changes of alignment.
(5) Use OFFSET(DescriptorSet, data) instead of sizeof(vk::DescriptorSetHeader) for calculating the offset of the data. The data field may have a different alignment to DescriptorSetHeader (in the future).
(6) Sanity check binding offsets end at getDescriptorSetDataSize() with an assert. These are calculated with two independent functions that can easily go out of sync.

Change-Id: I0bbc8355d32f80a72517e81bdf6d1d0d11b54845