Fix building with Clang 4.0
Clang 4.0.1 is less than two years old and still in widespread use.
It doesn't support some of the more recent diagnostic warnings.

For older verisons of CMake the CLANG_VERSION_STRING may not be defined,
so set it to CMAKE_CXX_COMPILER_VERSION after we've already conformed
that we're building with Clang.

Change-Id: I68c6270c2e90c50b1de1657e16e914c89b336952