Pass a separate sampler pointer to the sampling routine
Previously only the texture pointer was passed to the sampling routine,
containing some sampler data which is (still) wrongly assumed to come
from a combined image-view and sampler descriptor. This change will
allow splitting them up correctly.

Also refactor getImageSampler to not start the Reactor function
generation, moving it entirely into emitSamplerFunction so that the
latter only has to be passed the pointers to the descriptor state that
will affect the code that's generated.

Name all the parameters according to their actual type.

Clamp sampler LOD on construction so we don't have to deal with it any
more later.

Bug: b/129523279
Test: dEQP-VK.glsl.texture_functions.*
Test: dEQP-VK.spirv_assembly.instruction.graphics.image_sampler.*
Change-Id: Id3a9f12f379cf65741198b732fe387ec6e24dd86