LLVMReactor: Fix atomic loads and stores of non-scalar types.
The documentation and verifier both say the pointee needs to be a integer, pointer, or floating point type.
It looks like invalid atomic stores may be lowered to __atomic_load / __atomic_store via the AtomicExpandPass, but this comes much later in the transform pipeline, and before the verifier.

Bug b/127472316

Change-Id: Ieae5527ca5d65c890251c802baeaca6c5a9eacbb