Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The root of the problem here is that you're not tracking availability, you're tracking activeness.

Please explain. The way I read the spec, all queries start out as unavailable and MUST be unavailable up until a queue executes a command (set from a call to vkCmdBeginQuery) and can be set back to unavailable with a queue that executes a command (set from a call to vkCmdResetQueryPool).

The lifetime of a query seems to be:
unavailable -> begin -> available -> <get some results> -> reset -> unavailable -> begin -> available -> etc

If that's not how it works, then the spec using the word "unavailable" in the description of vkCmdBeginQuery/vkCmdResetQueryPool is confusing me.

Patch-set: 1
Reviewer: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>