Update patch set 2
Patch Set 2: Code-Review+1

I was just talking about this with Nicolas yesterday. This is fine. I would've liked to have some way to get notified of what extension an app is trying to use (and maybe get a large enough sample to collect some data on this in order to know which extensions are the most popular in the wild), but there doesn't seem to be any way to gather this data. Maybe there's already another way to get this kind of data somehow?

Patch-set: 2
Reviewer: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+1