Introduction of dawn's 32 bit safe code
In order to easily add 32 bit support, this cl adds some
(slightly modified) headers from the dawn project, which
already supports 32 bit platforms. This requires some
changes, since the Vk* types are now actual classes, so
we need to use the GetHandle() function where we used to
treat the Vk* object as a pointer and use the
Vk*::CreateFromHandle() function when we want to assign
a handle to an actual Vk* object.

This change should be noop.

Bug b/129979580

Change-Id: I54254929186584ec4544a1da5a7def7cf56e392e