Fix FP blending path to not make clamping assumptions
We had assumed in Context in various places that negative pixel values
would be clamped to zero, which is incorrect for floating-point color
attachments.

TODO: on the assumption that there is value in these blend optimizations
themselves, we should split the context blend state to be
per-attachment. This would give us the `independentBlend` optional
feature, and also allow us to keep the optimization for those
attachments which do have clamping in a mixed MRT scenario.

Test: dEQP-VK.pipeline.blend.format.r16*
Change-Id: I5434475aba5d21a8c8342b42cccc575b517f0900