System/Synchronization.hpp: Add the sw::TaskEvents interface
This is implemented by sw::WaitGroup, and vk::Fence.

The goal here is to abstract the underlying implementation of the task tracker. We currently (ab)use vk::Fence for queue tracking when a sw::WaitGroup could be used.

This also allows us to implement a task logger in the future that can show what's inflight at any time.

Bug: b/133127573
Change-Id: I0177b26e7577ffd5cb687a5105a69a22070a7ac6