System: Replace concurrent types with std versions where possible
Replace sw::MutexLock with std::mutex.
Replace use of sw::Thread with std::thread.

helgrind seems to take objection to the init Event.

AtomicInt remains as a wrapper around std::atomic<int> as it provides an 'operator=' and
'std::atomic& operator=( const std::atomic& )' is deleted.

This change is to reduce the amount of code I have to verify while looking for synchronization issues.

Bug: b/133127573
Change-Id: I8cbe38c918298ffafaf59fe25b0ff258670f9a99