Update patch set 1
Patch Set 1:

(1 comment)

So the destructors were originally removed to force us to use the destroy() functions, but this is fine as long as we use both. Since this is part of a template, the chances of forgetting it are slim enough, as long as noone starts allocating these objects themselves, outside of the proper API calls.

Patch-set: 1
CC: Gerrit User 5050 <5050@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>