Reimplement vk::Query using sw sync primitives.
Hide all the error-prone synchronization logic in the class, and expose a documented API.

This should have no impact on behavior - it was authored to make it harder to break things in the future.
That said, it appears to fix a whole bunch of flakes with the dEQP query tests.

Bug: b/133127573
Change-Id: I5c30b79b9b1cd36dba1fa2d3c34af0f5bd62772a