Update patch set 1
Patch Set 1:

After recommending to you that we comment this out, perhaps it might be more sensible to move these to a new vk-master-disabled.txt file or something.
The thought process here is that we'll likely update these lists, and it'll be easy to accidentally overwrite and forget any edits to these files.
Ideally, regres would allow you to take the unaltered file list and exclude test items from a 'disabled' list, so the original list doesn't have to be modified at all.
If the general consensus is that this is a good idea, I can knock this together. Until then, this works for me.

Patch-set: 1
Reviewer: Gerrit User 7183 <7183@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>