Fix taking the sampling instruction type into account
Our lookup key for generated sampling functions was only using the
sampler and image view descriptors into account. The same descriptors
can be used in various different sampling instructions, so we need to
also make that part of the lookup key.

This assumes the maximum number of these descriptors fits in a 16-bit
value (note that 0 is an invalid descriptor ID), which is validated with
a static assert.

Bug: b/129523279
Change-Id: I325aa5571191e4ffc69214c2d13d431d21316fed