Fix Fences on Windows
Adding a base class to the vk::Fence class on Windows caused
an issue with the "static inline Fence* Cast(VkFence object)"
function, which failed because of the extra space allocated
in the vtable at the beginning of the class (the "this"
pointer would always be off by 2 bytes).

Removed the base class and associated changes to fix Fences
on Windows.

Change-Id: I9aa876b18654e4eae851bf8e6e2f7304b27dda55