Fix Cast(vkFence) crashing on Windows
Fence has a second base class with virtual functions, but it's first base does not.
In "static VkResult Create(...), when assigning the handle via:

*outObject = *object;

since this is code run in the context of the first base class, the address returned
is offset by sizeof the vtable pointer from the second base class (8 bytes on x64).

Problem is that the Cast overload for Fence does a simple reinterpret_cast of this
stored pointer back to Fence*, without the offset being taken into account. The
fix is to simply first cast the stored pointer back to the first base class type, and
then static_cast it to the derived Fence type.

Change-Id: Ieb2856320e996a5ad5d9e69aad5e01cfaee1cd9e