Update patch set 1
Patch Set 1:

Pasting the more detailed explanation I wrote in the SwiftShader chat:


Okay, so here's the problem: vkFence derives from both Object<Fence, VkFence> and from sw::TaskEvents. The first base has no virtual functions, but the second base does. When we create a vkFence (or any object), we end up in "static VkResult Create(...)", which near the end executes the following to copy the pointer to the created object into the output VkT wrapper arg:

	*outObject = *object;

This line actually ends up calling an implicit cast operator to VkT ("operator VkT()") from the first base class, Object<Fence,...>, which does the following:

	operator VkT()
	{
		return reinterpret_cast<typename VkT::HandleType>(this);
	}

So, since vkFence's first base, Object<...>, has no vtable, but the second base, sw::TaskEvents, does, the 'this' we are returning here starts right after the vtable pointer for sw::TaskEvents, which is offset by 8 bytes. Indeed, we can see it in the disassembly for this assignment for vkFence:

	*outObject = *object;
00007FF862683D61  mov         rax,qword ptr [object]  
00007FF862683D66  add         rax,8  <------ this is the problem. Offset here to skip the vtable. We're basically slicing here.
00007FF862683D6A  lea         rdx,[rsp+40h]  
00007FF862683D6F  mov         rcx,rax  
00007FF862683D72  call        vk::Object<vk::Fence,VkWrapper<VkFence_T * __ptr64> >::operator VkWrapper<VkFence_T * __ptr64> (07FF8622AE0E9h)  
00007FF862683D77  mov         rax,qword ptr [rax]  
00007FF862683D7A  mov         rcx,qword ptr [outObject]  
00007FF862683D7F  mov         qword ptr [rcx],rax  

So we now return a pointer that is offset by 8 bytes. Later on, when we use Cast<VkFence>(), we do a simple reinterpret_cast of the address we stored earlier, but we are NOT offsetting by -8 bytes to get back to the start of the actual object in memory:

static inline Fence* Cast(VkFence object)
{
	return reinterpret_cast<Fence*>(object.get());
}

To fix this, we can simply modify the Cast function to first reinterpret_cast the address returned by object.get() to an Object<...>*, and then static_cast it to a VkFence:

static inline Fence* Cast(VkFence object)
{
	return static_cast<Fence*>(reinterpret_cast<Object<Fence, VkFence>*>(object.get()));
}

This is how we should do all our Cast functions, so that we can add multiple bases as we please. An alternative fix, but not a good one, would be to add a vtable to Object<>. For instance, we could simply add a defaulted virtual destructor to ObjectBase:

template<typename T, typename VkT>
class ObjectBase
{
public:
	virtual ~ObjectBase() = default;
...
};

This would work because now the start of the object in memory will be on the first vtable pointer, which now belongs to Object(Base). But this isn't great, because it adds a vtable pointer to all our objects for no real reason. The only advantage is that we get to keep all Cast implementations as they are without worrying about this problem arising again. Personally, though, I think we could easily write a generic cast template that all Cast calls can use; something like:

template <typename T, typename VkT>
static inline T* CastTo(VkT object)
{
	return static_cast<T*>(reinterpret_cast<Object<T, VkT>*>(object.get()));
}

And our Cast for Fence becomes:

static inline Fence* Cast(VkFence object)
{
	return CastTo<Fence>(object);
}

We could actually generate all the Casts with a single template if we had a compile-time mapping of type to VkType, but that could be done later.

Patch-set: 1
CC: Gerrit User 9669 <9669@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>